SCRUM-98 Refactor code, add reactivness, add some styling
