+++ frontend/src/stylesheets/GardenComponent.module.css
@@ -5,17 +5,6 @@
   align-items: center;
 }
 
-.CreateButton {
-  padding: 10px 20px;
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  border-radius: 6px;
-  font-weight: bold;
-  margin-bottom: 1rem;
-  cursor: pointer;
-}
-
 .SelectedGarden {
   width: 100%;
   display: flex;
@@ -34,9 +23,10 @@
   font-weight: bold;
   color: #2e7d32;
   background-color: #e8f5e9;
-  padding: 0.5rem 1.5rem;
+  padding: 0.5rem 5rem;
   border-radius: 8px;
   box-shadow: 0 1px 4px rgba(0, 0, 0, 0.1);
+  min-width: 5rem;
 }
 
 .MainLayout {
@@ -75,12 +65,3 @@
   background-color: #f3f3f3;
 }
 
-/* Right panel (menu) */
-.Menu {
-  flex: 0 0 250px;
-  background-color: beige;
-  padding: 1rem;
-  border-radius: 8px;
-  height: fit-content;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
